Feat: 이벤트 결제 완료 페이지 구현 및 퍼널로 페이지 관리하게 수정
